+++ src/main/java/com/appcenter/marketplace/domain/coupon/dto/res/CouponListResDto.java
@@ -2,21 +2,19 @@ package com.appcenter.marketplace.domain.coupon.dto.res;
 
 import lombok.Builder;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Getter
 public class CouponListResDto {
-    private final List<CouponResDto> couponList;
+    private final List<CouponMemberResDto> couponList;
 
     @Builder
-    public CouponListResDto(List<CouponResDto> couponList){
+    public CouponListResDto(List<CouponMemberResDto> couponList){
         this.couponList = couponList;
     }
 
-    public static CouponListResDto toDto(List<CouponResDto> couponList){
+    public static CouponListResDto toDto(List<CouponMemberResDto> couponList){
        return CouponListResDto.builder().couponList(couponList).build();
     }
 }
