+++ src/main/java/com/appcenter/marketplace/domain/member_coupon/repository/MemberCouponRepositoryCustom.java
@@ -1,5 +1,5 @@
 package com.appcenter.marketplace.domain.member_coupon.repository;
 
 public interface MemberCouponRepositoryCustom {
-    public boolean existCouponByMemberId(Long memberId, Long couponId);
+    boolean existCouponByMemberId(Long memberId, Long couponId);
 }
